fix: Remove bfj & fs calls from core logic
