+++ main.js
@@ -1,10 +1,7 @@
-
-
 export * as groth16 from "./src/groth16.js";
 export * as powersOfTau from "./src/powersoftau.js";
 export * as r1cs from "./src/r1cs.js";
 export * as wtns from "./src/wtns.js";
-export * as wtnsCmds from "./src/cmds/wtns_cmds.js";
 export * as zKey from "./src/zkey.js";
 export * as plonk from "./src/plonk.js";
-export * as fflonk from "./src/cmds/fflonk_cmds.js";
+export * as fflonk from "./src/fflonk.js";
