--- src/cmds/fflonk_cmds.js
@@ -1,25 +1,0 @@
-/*
-    Copyright 2022 iden3 association.
-
-    This file is part of snarkJS.
-
-    snarkJS is a free software: you can redistribute it and/or modify it
-    under the terms of the GNU General Public License as published by
-    the Free Software Foundation, either version 3 of the License, or
-    (at your option) any later version.
-
-    snarkJS is distributed in the hope that it will be useful, but WITHOUT
-    ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
-    or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public
-    License for more details.
-
-    You should have received a copy of the GNU General Public License
-    along with snarkJS. If not, see <https://www.gnu.org/licenses/>.
-*/
-
-export {fflonkSetupCmd} from "./fflonk_setup_cmd.js";
-export {fflonkProveCmd} from "./fflonk_prove_cmd.js";
-export {fflonkFullProveCmd} from "./fflonk_full_prove_cmd.js";
-export {fflonkVerifyCmd} from "./fflonk_verify_cmd.js";
-export {fflonkExportSolidityVerifierCmd} from "./fflonk_export_solidity_verifier_cmd.js";
-export {fflonkExportCallDataCmd} from "./fflonk_export_calldata_cmd.js";
