--- src/cmds/fflonk_setup_cmd.js
@@ -1,22 +1,0 @@
-/*
-    This file is part of snarkjs.
-
-    snarkjs is a free software: you can redistribute it and/or
-    modify it under the terms of the GNU General Public License as published by the
-    Free Software Foundation, either version 3 of the License, or (at your option)
-    any later version.
-
-    snarkjs is distributed in the hope that it will be useful,
-    but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
-    or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for
-    more details.
-
-    You should have received a copy of the GNU General Public License along with
-    snarkjs. If not, see <https://www.gnu.org/licenses/>.
-*/
-
-import fflonkSetup from "../fflonk_setup.js";
-
-export async function fflonkSetupCmd(r1csFilename, ptauFilename, zkeyFilename, logger) {
-    return fflonkSetup(r1csFilename, ptauFilename, zkeyFilename, logger);
-}
