+++ src/fflonk_constants.js
@@ -1,0 +1,44 @@
+/*
+    Copyright 2022 iden3 association.
+
+    This file is part of snarkjs.
+
+    snarkjs is a free software: you can redistribute it and/or
+    modify it under the terms of the GNU General Public License as published by the
+    Free Software Foundation, either version 3 of the License, or (at your option)
+    any later version.
+
+    snarkjs is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
+    or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for
+    more details.
+
+    You should have received a copy of the GNU General Public License along with
+    snarkjs. If not, see <https://www.gnu.org/licenses/>.
+*/
+
+// FFlonk constants
+export const FF_T_POL_DEG_MIN = 3;
+
+// FFlonk A
+export const A = 12;
+
+// ZKEY constants
+export const ZKEY_FF_NSECTIONS = 17;
+
+export const ZKEY_FF_HEADER_SECTION = 2;
+export const ZKEY_FF_ADDITIONS_SECTION = 3;
+export const ZKEY_FF_A_MAP_SECTION = 4;
+export const ZKEY_FF_B_MAP_SECTION = 5;
+export const ZKEY_FF_C_MAP_SECTION = 6;
+export const ZKEY_FF_QL_SECTION = 7;
+export const ZKEY_FF_QR_SECTION = 8;
+export const ZKEY_FF_QM_SECTION = 9;
+export const ZKEY_FF_QO_SECTION = 10;
+export const ZKEY_FF_QC_SECTION = 11;
+export const ZKEY_FF_SIGMA1_SECTION = 12;
+export const ZKEY_FF_SIGMA2_SECTION = 13;
+export const ZKEY_FF_SIGMA3_SECTION = 14;
+export const ZKEY_FF_LAGRANGE_SECTION = 15;
+export const ZKEY_FF_PTAU_SECTION = 16;
+export const ZKEY_FF_C0_SECTION = 17;
