+++ src/fflonk_export_calldata.js
@@ -33,7 +33,7 @@ function p256(n) {
     return nstr;
 }
 
-export default async function fflonkExportCallData(_pub, _proof, logger) {
+export default async function fflonkExportCallData(_pub, _proof) {
     const proof = unstringifyBigInts(_proof);
     const pub = unstringifyBigInts(_pub);
