+++ src/wtns.js
@@ -20,3 +20,4 @@
 export {default as calculate} from "./wtns_calculate.js";
 export {default as debug} from "./wtns_debug.js";
 export {default as exportJson} from "./wtns_export_json.js";
+export {default as check} from "./wtns_check.js";
