+++ src/zkey_export_solidityverifier.js
@@ -1,8 +1,7 @@
-import * as fastFile from "fastfile";
 import ejs from "ejs";
 
 import exportVerificationKey from "./zkey_export_verificationkey.js";
-import {fflonkExportSolidityVerifierCmd} from "./cmds/fflonk_cmds.js";
+import fflonkExportSolidityVerifierCmd from "./fflonk_export_solidity_verifier.js";
 // Not ready yet
 // module.exports.generateVerifier_kimleeoh = generateVerifier_kimleeoh;
